breaking_change: change of package manager
The package manager "packer" was changed to "lazy" The order
of requiring the files in root init.lua must be changed. The plugins
file must be first because the lazy configuration sets
vim.opt.mapleader options and it must be set right after lazy setup.

The loader code is dropped because of compatibility reasons.
This change is not set to be permanent.

Signed-off-by: Filip Lobpreis <f.lobpreis@gmail.com>