fixup! fix: keymap: pasting in visual mode
There is actually no easy fix for this. Even if you check for the
'> mark position. The position is updated after one operation and
thus there is no way to get it right on the first try. Just make
the pasing manually. If you want to paste a visual selection to the
end of a line to a visual selection just press P otherwise press p.